refactor to LockManager class
